OOP validate hook
