+++ Tests/scripts/hook_validations/secrets.py
@@ -1,5 +1,3 @@
-from __future__ import absolute_import
-
 import io
 import os
 import re
